facial-recognition-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
deepface==0.0.92
tensorflow==2.15.0
tf-keras==2.15.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.3
aiofiles==23.2.1
motor==3.3.2
pymongo==4.6.1
